requirements.txt
alembic==1.3.3
alipay-sdk-python==3.3.398
amqp==2.5.2
backcall==0.1.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
decorator==4.4.1
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.0
kombu==4.6.7
Mako==1.1.1
MarkupSafe==1.1.1
monotonic==1.5
mysql-connector==2.2.9
mysql-connector-python==8.0.19
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.8
pycrypto==2.6.1
Pygments==2.5.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
qiniu==7.2.6
redis==3.4.1
requests==2.22.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.13
traitlets==4.3.3
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
Werkzeug==0.16.1
WTForms==2.2.1
xmltodict==0.12.0
xmltojson==0.2.0
zipp==2.1.0
